pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aldemarvin-extractor"
version = "1.0.0"
description = "Sistema de extração de texto de imagens com tradução e organização em livros/artigos"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Aldemarvin"}
]
dependencies = [
    "tinydb>=4.8.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.4.0",
    "deep-translator>=1.11.4",
    "fpdf2>=2.7.9",
    "pyperclip>=1.9.0",
]

[project.scripts]
aldemarvin = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
